pyproject.toml
[tool.poetry]
name = "primalscheme"
version = "2.0.0rc1"
description = ""
authors = ["Andrew Smith <andysmith.d@gmail.com>", "Josh Quick <j.quick@bham.ac.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.1"
primer3-py = "^0.6.1"
parasail = "^1.2.4"
biopython = "^1.79"
Jinja2 = "^3.0.3"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.812"
black = "^21.5b1"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
primalscheme = 'primalscheme.main:main'